Web: refactor the huge `hubApiService`
- Moved all general API helpers into a separate `apiHelpers` service.
- Added a new `withPaginationParams` helper as part of this, to DRY up how pagination params are provided to API calls.
- Made the default "per page" value an app global constant.
- Moved all the request builders into a separate `apiRequestBuilders` service.
